+++ packages/create-dawn-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-dawn-ai-app
 
+## 0.8.2
+
+### Patch Changes
+
+- @dawn-ai/devkit@0.8.2
+
 ## 0.8.1
 
 ### Patch Changes