+++ packages/sqlite-storage/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @dawn-ai/sqlite-storage
 
+## 0.8.2
+
 ## 0.8.1
 
 ### Patch Changes